Add failsafe option to send notification
In some environments – e.g. Webstorm/Intellij tasks – the streams never receive the 'data' or 'end' events, so notifications are not shown and the notifier process never ends.

This option will force the notification to send after n number of milliseconds.

Also changed the callback to always exit the process, not just when a message is returned from the notifier.